+++ src/app/components/editora/editora-list/editora-list.component.ts
@@ -7,6 +7,8 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { MatTableModule } from '@angular/material/table';
 import { RouterModule } from '@angular/router';
+import { MatDialog } from '@angular/material/dialog'; // Importa MatDialog
+import { ConfirmationDialogComponent } from '../../dialog/confirmation-dialog/confirmation-dialog.component'; // Importa o componente de confirmação
 
 @Component({
   selector: 'app-editora-list',
@@ -19,7 +21,7 @@ export class EditoraListComponent implements OnInit {
   editoras: Editora[] = [];
   displayedColumns: string[] = ['id', 'nome', 'email', 'endereco', 'telefone', 'acao'];
 
-  constructor(private editoraService: EditoraService) {}
+  constructor(private editoraService: EditoraService, private dialog: MatDialog) {} // Injeta MatDialog
 
   ngOnInit(): void {
     this.editoraService.findAll().subscribe(
@@ -28,12 +30,22 @@ export class EditoraListComponent implements OnInit {
   }
 
   excluir(editora: Editora): void {
-    this.editoraService.delete(editora).subscribe({
-      next: () => {
-        this.editoras = this.editoras.filter(e => e.id !== editora.id);
-      },
-      error: (err) => {
-        console.error("Erro ao tentar excluir a editora", err);
+    // Abre o diálogo de confirmação
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: { message: 'Deseja realmente excluir esta editora?' }
+    });
+
+    // Processa a resposta do diálogo
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.editoraService.delete(editora).subscribe({
+          next: () => {
+            this.editoras = this.editoras.filter(e => e.id !== editora.id);
+          },
+          error: (err) => {
+            console.error("Erro ao tentar excluir a editora", err);
+          }
+        });
       }
     });
   }
