+++ src/app/components/genero/genero-form/genero-form.component.ts
@@ -10,31 +10,43 @@ import { MatToolbarModule } from '@angular/material/toolbar';
 import { ActivatedRoute, Router, RouterModule } from '@angular/router';
 import { GeneroService } from '../../../services/genero.service';
 import { Genero } from '../../../models/genero.model';
-
+import { MatDialog } from '@angular/material/dialog'; // Importa MatDialog
+import { ConfirmationDialogComponent } from '../../dialog/confirmation-dialog/confirmation-dialog.component'; // Importa o componente de diálogo de confirmação
 
 @Component({
   selector: 'app-genero-form',
   standalone: true,
-  imports: [ReactiveFormsModule, MatCardModule, MatFormFieldModule,
-    MatButtonModule, NgIf, MatInputModule, RouterModule, MatTableModule, MatToolbarModule],
+  imports: [
+    ReactiveFormsModule, 
+    MatCardModule, 
+    MatFormFieldModule, 
+    MatButtonModule, 
+    NgIf, 
+    MatInputModule, 
+    RouterModule, 
+    MatTableModule, 
+    MatToolbarModule
+  ],
   templateUrl: './genero-form.component.html',
-  styleUrl: './genero-form.component.css'
+  styleUrls: ['./genero-form.component.css'] // Corrigido para 'styleUrls'
 })
 export class GeneroFormComponent {
   formGroup: FormGroup;
 
-  constructor(private formBuilder: FormBuilder,
+  constructor(
+    private formBuilder: FormBuilder,
     private generoService: GeneroService,
     private router: Router,
-    private activatedRoute: ActivatedRoute) {
+    private activatedRoute: ActivatedRoute,
+    private dialog: MatDialog // Injeta o MatDialog
+  ) {
+    const genero: Genero = this.activatedRoute.snapshot.data['genero'];
 
-      const genero: Genero = this.activatedRoute.snapshot.data['genero'];
-
-      this.formGroup = this.formBuilder.group({
-        id: [(genero && genero.id) ? genero.id : null],
-        nome: [(genero && genero.nome) ? genero.nome : '', Validators.required],
-        descricao: [(genero && genero.descricao) ? genero.descricao : '', Validators.required]
-    })
+    this.formGroup = this.formBuilder.group({
+      id: [(genero && genero.id) ? genero.id : null],
+      nome: [(genero && genero.nome) ? genero.nome : '', Validators.required],
+      descricao: [(genero && genero.descricao) ? genero.descricao : '', Validators.required]
+    });
   }
 
   salvar() {
@@ -43,7 +55,7 @@ export class GeneroFormComponent {
       if (genero.id == null) {
         this.generoService.insert(genero).subscribe({
           next: (generoCadastrado) => {
-            this.router.navigateByUrl('/generos')
+            this.router.navigateByUrl('/generos');
           },
           error: (errorResponse) => {
             console.log('Erro ao salvar' + JSON.stringify(errorResponse));
@@ -66,15 +78,28 @@ export class GeneroFormComponent {
     if (this.formGroup.valid) {
       const genero = this.formGroup.value;
       if (genero.id != null) {
-        this.generoService.delete(genero).subscribe({
-          next: () => {
-            this.router.navigateByUrl('/generos');
-          },
-          error: (err) => {
-            console.log('Erro ao excluir' + JSON.stringify(err));
+        // Abre o diálogo de confirmação
+        const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+          data: { message: 'Deseja realmente excluir este Gênero? Não será possível reverter.' }
+        });
+
+        dialogRef.afterClosed().subscribe(result => {
+          if (result) {
+            this.generoService.delete(genero).subscribe({
+              next: () => {
+                this.router.navigateByUrl('/generos');
+              },
+              error: (err) => {
+                console.log('Erro ao excluir' + JSON.stringify(err));
+              }
+            });
           }
         });
       }
     }
+  }
+
+  cancelar() {
+    this.router.navigateByUrl('/generos');
   }
 }
